requirements.txt
pandas
numpy
faker
tqdm
transformers
torch
seaborn
matplotlib
xgboost
plotly
streamlit
imbalanced-learn
streamlit-echarts
wordcloud
scikit-learn
flask

# Azure SDK packages
azure-storage-blob>=12.19.0
azure-ai-textanalytics>=5.3.0
azure-identity>=1.15.0
azure-ai-ml>=1.12.0
azure-functions>=1.18.0
azure-core>=1.29.0
python-dotenv>=1.0.0